// apps/api/src/modules/agents/domain/errors/agent-errors.ts
export enum AgentErrorCode {
  AGENT_NOT_FOUND = 'AGENT_NOT_FOUND',
  AGENT_VERSION_NOT_FOUND = 'AGENT_VERSION_NOT_FOUND',
  AGENT_DISABLED = 'AGENT_DISABLED',
  AGENT_DEPRECATED = 'AGENT_DEPRECATED',
  AGENT_UNAVAILABLE = 'AGENT_UNAVAILABLE',
  AGENT_MISCONFIGURED = 'AGENT_MISCONFIGURED',
  AGENT_POLICY_DENIED = 'AGENT_POLICY_DENIED',
  AGENT_AUTHENTICATION_REQUIRED = 'AGENT_AUTHENTICATION_REQUIRED',
  AGENT_USER_CONTEXT_REQUIRED = 'AGENT_USER_CONTEXT_REQUIRED',
  AGENT_INPUT_INVALID = 'AGENT_INPUT_INVALID',
  AGENT_CONTEXT_INVALID = 'AGENT_CONTEXT_INVALID',
  AGENT_CONTEXT_STALE = 'AGENT_CONTEXT_STALE',
  AGENT_PROMPT_NOT_FOUND = 'AGENT_PROMPT_NOT_FOUND',
  AGENT_PROMPT_INVALID = 'AGENT_PROMPT_INVALID',
  AGENT_MODEL_UNAVAILABLE = 'AGENT_MODEL_UNAVAILABLE',
  AGENT_PROVIDER_UNAVAILABLE = 'AGENT_PROVIDER_UNAVAILABLE',
  AGENT_TOOL_UNAVAILABLE = 'AGENT_TOOL_UNAVAILABLE',
  AGENT_BUDGET_EXCEEDED = 'AGENT_BUDGET_EXCEEDED',
  AGENT_QUOTA_EXCEEDED = 'AGENT_QUOTA_EXCEEDED',
  AGENT_CONCURRENCY_EXCEEDED = 'AGENT_CONCURRENCY_EXCEEDED',
  AGENT_TIMEOUT = 'AGENT_TIMEOUT',
  AGENT_CANCELLED = 'AGENT_CANCELLED',
  AGENT_OUTPUT_INVALID = 'AGENT_OUTPUT_INVALID',
  AGENT_RETRY_EXHAUSTED = 'AGENT_RETRY_EXHAUSTED',
  AGENT_DUPLICATE_RUN = 'AGENT_DUPLICATE_RUN',
  AGENT_STATE_TRANSITION_INVALID = 'AGENT_STATE_TRANSITION_INVALID',
  AGENT_SECRET_DETECTED = 'AGENT_SECRET_DETECTED',
  AGENT_EXECUTION_FAILED = 'AGENT_EXECUTION_FAILED',
  AGENT_UNKNOWN_ERROR = 'AGENT_UNKNOWN_ERROR'
}

export class AgentError extends Error {
  public readonly code: AgentErrorCode;
  public readonly safeMessage: string;
  public readonly retryable: boolean;

  constructor(code: AgentErrorCode, safeMessage: string, retryable: boolean = false) {
    super(safeMessage);
    this.name = 'AgentError';
    this.code = code;
    this.safeMessage = safeMessage;
    this.retryable = retryable;
  }
}